requirements.txt
django==2.0.6
djangorestframework

